Consolidate and conform to pylint standard
